Refactor: review and move timeout pieces back
